Add {{markdown-render}} helper
Closes https://github.com/greenfieldhq/ember-markdown-it/issues/1

This helper can be used in templates in order to render markdown as
HTML. It returns an `Ember.Handlebars.SafeString`, which may not always
be what you want to allow, but we're relying on markdown-it to provide
sanitization of the string for us.

In the future, I'd like to make this configurable.